split UserExerciseRepository
