+++ app/Structures/AuthenticatedUserExerciseRepository.php
@@ -1,0 +1,88 @@
+<?php
+
+namespace App\Structures;
+
+use App\Models\User;
+use Illuminate\Database\Query\Builder;
+use Illuminate\Database\Query\JoinClause;
+use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\DB;
+
+class AuthenticatedUserExerciseRepository implements AuthenticatedUserExerciseRepositoryInterface
+{
+    /**
+     * @param User $user
+     * @param int  $exerciseId
+     * @return UserExercise
+     * @throws \Exception
+     */
+    public function fetchUserExerciseOfExercise(User $user, int $exerciseId): UserExercise
+    {
+        $result = DB::table('exercises AS e')
+            ->select([
+                'e.id AS exercise_id',
+                'e.lesson_id AS lesson_id',
+                'e.question AS question',
+                'e.answer AS answer',
+                DB::raw('COALESCE(er.number_of_good_answers, 0) AS number_of_good_answers'),
+                DB::raw('COALESCE(er.number_of_good_answers_today, 0) AS number_of_good_answers_today'),
+                'er.latest_good_answer AS latest_good_answer',
+                DB::raw('COALESCE(er.number_of_bad_answers, 0) AS number_of_bad_answers'),
+                DB::raw('COALESCE(er.number_of_bad_answers_today, 0) AS number_of_bad_answers_today'),
+                'er.latest_bad_answer AS latest_bad_answer',
+                DB::raw('COALESCE(er.percent_of_good_answers, 0) AS percent_of_good_answers'),
+            ])
+            ->leftJoin('exercise_results AS er', function (JoinClause $joinClause) use ($user) {
+                $joinClause
+                    ->on('er.exercise_id', '=', 'e.id')
+                    ->where('er.user_id', '=', $user->id);
+            })
+            ->where('e.id', '=', $exerciseId)
+            ->first();
+
+        if (!$result) {
+            throw new \Exception('Exercise does not exist: '.$exerciseId);
+        }
+
+        return new UserExercise($result);
+    }
+
+    /**
+     * @param User   $user
+     * @param string $phrase
+     * @return Collection|UserExercise[]
+     */
+    public function fetchUserExercisesWithPhrase(User $user, string $phrase): Collection
+    {
+        return DB::table('exercises AS e')
+            ->select([
+                'e.id AS exercise_id',
+                'e.lesson_id AS lesson_id',
+                'l.name AS lesson_name',
+                'e.question AS question',
+                'e.answer AS answer',
+                DB::raw('COALESCE(er.number_of_good_answers, 0) AS number_of_good_answers'),
+                DB::raw('COALESCE(er.number_of_good_answers_today, 0) AS number_of_good_answers_today'),
+                'er.latest_good_answer AS latest_good_answer',
+                DB::raw('COALESCE(er.number_of_bad_answers, 0) AS number_of_bad_answers'),
+                DB::raw('COALESCE(er.number_of_bad_answers_today, 0) AS number_of_bad_answers_today'),
+                'er.latest_bad_answer AS latest_bad_answer',
+                DB::raw('COALESCE(er.percent_of_good_answers, 0) AS percent_of_good_answers'),
+            ])
+            ->join('lessons AS l', function (JoinClause $join) use ($user) {
+                $join->on('l.id', '=', 'e.lesson_id')
+                    ->where('l.owner_id', '=', $user->id);
+            })
+            ->where(function (Builder $builder) use ($phrase) {
+                $builder->where('question', 'like', '%'.$phrase.'%')
+                    ->orWhere('answer', 'like', '%'.$phrase.'%');
+            })
+            ->leftJoin('exercise_results AS er', function (JoinClause $joinClause) use ($user) {
+                $joinClause
+                    ->on('er.exercise_id', '=', 'e.id')
+                    ->where('er.user_id', '=', $user->id);
+            })
+            ->get()
+            ->mapInto(UserExercise::class);
+    }
+}
